refactor endpoint sourcecode structure
