+++ components/buildless-serverless/internal/endpoint/response.go
@@ -6,18 +6,15 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/kyma-project/serverless/components/buildless-serverless/internal/endpoint/types"
 	"github.com/pkg/errors"
 )
-
-type ErrorResponse struct {
-	Error string `json:"error"`
-}
 
 func (s *Server) writeErrorResponse(w http.ResponseWriter, status int, respErr error) {
 	headerStatus := status
 
 	buf := bytes.NewBuffer([]byte{})
-	err := json.NewEncoder(buf).Encode(ErrorResponse{Error: respErr.Error()})
+	err := json.NewEncoder(buf).Encode(types.ErrorResponse{Error: respErr.Error()})
 	if err != nil {
 		// If encoding fails, we can't do much, so we log the error and return a generic error response
 		s.log.Errorf("failed to encode error response: %v", err)
@@ -31,18 +28,9 @@ func (s *Server) writeErrorResponse(w http.ResponseWriter, status int, respErr e
 	fmt.Fprint(w, buf.String())
 }
 
-type FileResponse struct {
-	Name string `json:"name"`
-	Data string `json:"data"` // base64 encoded file content
-}
-
-type FilesListResponse struct {
-	Files []FileResponse `json:"files"`
-}
-
-func (s *Server) writeFilesListResponse(w http.ResponseWriter, data []FileResponse) {
+func (s *Server) writeFilesListResponse(w http.ResponseWriter, data []types.FileResponse) {
 	buf := bytes.NewBuffer([]byte{})
-	err := json.NewEncoder(buf).Encode(FilesListResponse{Files: data})
+	err := json.NewEncoder(buf).Encode(types.FilesListResponse{Files: data})
 	if err != nil {
 		s.writeErrorResponse(w, http.StatusInternalServerError, errors.Wrap(err, "failed to encode response"))
 		return
